// src/app/models/user-model.ts
export class User {
    UserId: number;
    DisplayName: string;
    UserName: string;
    MobileNo: string;
    EmailId: string;
    UPassword: string;
    isActive: number;
    RoleId: number;
    ProfileImage: string;
    isDefault: number;
    ResetOnNextLogin: number;
    SendMail: number;
    RefId: number;
    ReportTo: number;
    isPrimary?: number;
    Country?: number;
    ProvinceId?:any;
    SkypeId?: string;
    CountryId?: number;
    State?: string;
    City?: string;
    ZipCode?: string;
    InstEmail?: string;
    InstUser?: string;
    InstPassword?: string;
    InstSMTPServer?: string;
    InstSMTPPort?: string;
    InstSSL?: boolean;
    OfficeEmail?: string;
    OfficeUser?: string;
    OfficePassword?: string;
    OfficeSMTPServer?: string;
    OfficeSMTPPort?: string;
    OfficeSSL?: boolean;
    DepartmentId?: number;
    CampusId?: any;
}